[#44] Refactor: useQuery, useMutation 함수 분리
