+++ components/calendar-monthly/calendar-todos.tsx
@@ -1,28 +1,13 @@
 import { useCalendarContext } from '../calendar-compound/calendar';
-import { fetchDailyTodos, postRefreshToken } from '@/api/calendar/request';
-import { useQuery } from '@tanstack/react-query';
 import CalendarTodo from './calendar-todo';
+import useMonthlyTodosQuery from '@/hooks/queries/calendar/use-monthly-todos-query';
 
 export default function CalendarTodos() {
   const calendarContext = useCalendarContext();
   const { year, month, date } = calendarContext;
   const yearMonthDate = [year, (month + 1).toString().padStart(2, '0'), date.toString().padStart(2, '0')].join('-');
 
-  const {
-    data: dailyTodos,
-    error,
-    isPending,
-    isError,
-  } = useQuery({
-    queryKey: ['dailyTodos', yearMonthDate],
-    queryFn: () => fetchDailyTodos(yearMonthDate),
-    placeholderData: (prevDailyTodos) => prevDailyTodos,
-    retry: (count) => {
-      if (count > 1) return false;
-      if (count === 0) postRefreshToken();
-      return true;
-    },
-  });
+  const { data: dailyTodos, error, isPending, isError } = useMonthlyTodosQuery(yearMonthDate);
 
   if (isPending) return <span>loading</span>;
   if (isError) return <span>Error: {error.message}</span>;
